+++ src/RcppExports.cpp
@@ -56,8 +56,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // tableGT
-List tableGT(std::string vcffile, std::string region, std::string samples, std::string format, const std::vector<std::string>& ids, double qualval, bool pass, bool INFO, bool snps, bool indels, bool multiallelics, bool multisnps);
-RcppExport SEXP _vcfppR_tableGT(SEXP vcffileSEXP, SEXP regionSEXP, SEXP samplesSEXP, SEXP formatSEXP, SEXP idsSEXP, SEXP qualvalSEXP, SEXP passSEXP, SEXP INFOSEXP, SEXP snpsSEXP, SEXP indelsSEXP, SEXP multiallelicsSEXP, SEXP multisnpsSEXP) {
+List tableGT(std::string vcffile, std::string region, std::string samples, std::string format, const std::vector<std::string>& ids, double qualval, bool pass, bool INFO, bool snps, bool indels, bool multiallelics, bool multisnps, bool svs);
+RcppExport SEXP _vcfppR_tableGT(SEXP vcffileSEXP, SEXP regionSEXP, SEXP samplesSEXP, SEXP formatSEXP, SEXP idsSEXP, SEXP qualvalSEXP, SEXP passSEXP, SEXP INFOSEXP, SEXP snpsSEXP, SEXP indelsSEXP, SEXP multiallelicsSEXP, SEXP multisnpsSEXP, SEXP svsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -73,13 +73,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type indels(indelsSEXP);
     Rcpp::traits::input_parameter< bool >::type multiallelics(multiallelicsSEXP);
     Rcpp::traits::input_parameter< bool >::type multisnps(multisnpsSEXP);
-    rcpp_result_gen = Rcpp::wrap(tableGT(vcffile, region, samples, format, ids, qualval, pass, INFO, snps, indels, multiallelics, multisnps));
+    Rcpp::traits::input_parameter< bool >::type svs(svsSEXP);
+    rcpp_result_gen = Rcpp::wrap(tableGT(vcffile, region, samples, format, ids, qualval, pass, INFO, snps, indels, multiallelics, multisnps, svs));
     return rcpp_result_gen;
 END_RCPP
 }
 // tableFormat
-List tableFormat(std::string vcffile, std::string region, std::string samples, std::string format, const std::vector<std::string>& ids, double qualval, bool pass, bool INFO, bool snps, bool indels, bool multiallelics, bool multisnps);
-RcppExport SEXP _vcfppR_tableFormat(SEXP vcffileSEXP, SEXP regionSEXP, SEXP samplesSEXP, SEXP formatSEXP, SEXP idsSEXP, SEXP qualvalSEXP, SEXP passSEXP, SEXP INFOSEXP, SEXP snpsSEXP, SEXP indelsSEXP, SEXP multiallelicsSEXP, SEXP multisnpsSEXP) {
+List tableFormat(std::string vcffile, std::string region, std::string samples, std::string format, const std::vector<std::string>& ids, double qualval, bool pass, bool INFO, bool snps, bool indels, bool multiallelics, bool multisnps, bool svs);
+RcppExport SEXP _vcfppR_tableFormat(SEXP vcffileSEXP, SEXP regionSEXP, SEXP samplesSEXP, SEXP formatSEXP, SEXP idsSEXP, SEXP qualvalSEXP, SEXP passSEXP, SEXP INFOSEXP, SEXP snpsSEXP, SEXP indelsSEXP, SEXP multiallelicsSEXP, SEXP multisnpsSEXP, SEXP svsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -95,7 +96,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type indels(indelsSEXP);
     Rcpp::traits::input_parameter< bool >::type multiallelics(multiallelicsSEXP);
     Rcpp::traits::input_parameter< bool >::type multisnps(multisnpsSEXP);
-    rcpp_result_gen = Rcpp::wrap(tableFormat(vcffile, region, samples, format, ids, qualval, pass, INFO, snps, indels, multiallelics, multisnps));
+    Rcpp::traits::input_parameter< bool >::type svs(svsSEXP);
+    rcpp_result_gen = Rcpp::wrap(tableFormat(vcffile, region, samples, format, ids, qualval, pass, INFO, snps, indels, multiallelics, multisnps, svs));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -107,8 +109,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_vcfppR_heterozygosity", (DL_FUNC) &_vcfppR_heterozygosity, 5},
     {"_vcfppR_summaryVariants", (DL_FUNC) &_vcfppR_summaryVariants, 5},
     {"_vcfppR_summarySVs", (DL_FUNC) &_vcfppR_summarySVs, 5},
-    {"_vcfppR_tableGT", (DL_FUNC) &_vcfppR_tableGT, 12},
-    {"_vcfppR_tableFormat", (DL_FUNC) &_vcfppR_tableFormat, 12},
+    {"_vcfppR_tableGT", (DL_FUNC) &_vcfppR_tableGT, 13},
+    {"_vcfppR_tableFormat", (DL_FUNC) &_vcfppR_tableFormat, 13},
     {"_rcpp_module_boot_vcfreader", (DL_FUNC) &_rcpp_module_boot_vcfreader, 0},
     {"_rcpp_module_boot_vcfwriter", (DL_FUNC) &_rcpp_module_boot_vcfwriter, 0},
     {NULL, NULL, 0}